(#106) Criando a tablema Matricula e unificando as models - https://github.com/fga-eps-mds/2022.2-Amis-Doc/issues/106
